pyproject.toml
[tool.poetry]
name = "self"
version = "0.1.0"
description = ""
authors = ["Михаил Каллио <mikhailkallio1995@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.111.0"}
sqlalchemy = {extras = ["all"], version = "^2.0.31"}
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
greenlet = "^3.0.3"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
